requirements.txt
# Core
google-genai>=1.0.0
google-cloud-storage>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0

# Testing
pytest>=7.0.0

# Optional (for advanced features)
# mermaid-py  # Architecture diagram generation
# moviepy     # Alternative video editing
